schema generator: merge common.py back into validation.py
on second thoughts we're not going to be able to sensibly re-use any of the
common schema generation tools, so this common.py isn't going to be useful
