+++ schema_generator/validation.py
@@ -2,7 +2,6 @@ import os
 import re
 import json
 from dmcontent import ContentLoader
-from .common import merge_schemas, empty_schema
 
 
 MANIFESTS = {
@@ -95,6 +94,35 @@ def load_questions(schema_type, framework_slug, lot_slug):
         dynamic=False
     )
     return {q['id']: q for q in sum((s.questions for s in manifest.sections), [])}
+
+
+def merge_schemas(a, b):
+    if not (isinstance(a, dict) and isinstance(b, dict)):
+        raise TypeError("Error merging unsupported types '{}' and '{}'".format(
+            type(a).__name__, type(b).__name__
+        ))
+
+    result = a.copy()
+    for key, val in b.items():
+        if isinstance(result.get(key), dict):
+            result[key] = merge_schemas(a[key], val)
+        elif isinstance(result.get(key), list):
+            result[key] = result[key] + val
+        else:
+            result[key] = val
+
+    return result
+
+
+def empty_schema(schema_name):
+    return {
+        "title": "{} Schema".format(schema_name),
+        "$schema": "http://json-schema.org/schema#",
+        "type": "object",
+        "additionalProperties": False,
+        "properties": {},
+        "required": [],
+    }
 
 
 def drop_non_schema_questions(questions):
